Random Forest Practise Problems
Random Forest is a machine learning algorithm that uses ensemble learning method for classification, regression and other tasks. It operates by constructing a multitude of decision trees at training time and outputting the class that is the mode of the classes (classification) or mean prediction (regression) of the individual trees. It corrects for decision trees' habit of overfitting to their training set. Random Forests are flexible, easy to use, and provide high accuracy even with default parameters.